修改接口

+++ thrift/Comment.thrift
@@ -2,13 +2,13 @@ namespace java com.blog.thrift.domain
 namespace php common.thrift.gen.domain
 
 struct Comment {
-    1: i32 id,
-    2: i32 post_id,
-    3: i32 parent_id,
-    4: i8 status,
+    1: int id,
+    2: int post_id,
+    3: int parent_id,
+    4: int status,
     5: string content,
-    6: i32 create_time,
-    7: i32 update_time,
-    8: i32 author_id,
-    9: i32 approve_num,
+    6: int create_time,
+    7: int update_time,
+    8: int author_id,
+    9: int approve_num,
 }
