+++ thrift/Tag.thrift
@@ -2,10 +2,10 @@ namespace java com.blog.thrift.domain
 namespace php common.thrift.gen.domain
 
 struct Tag {
-    1: i32 id,
+    1: int id,
     2: string name,
-    3: i32 frequency,
-    4: i32 create_time,
-    5: i32 update_time,
-    6: i8 status,
+    3: int frequency,
+    4: int create_time,
+    5: int update_time,
+    6: int status,
 }
